move user post popup to new component
