+++ src/components/UserPosts/UserPosts.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react'
+import SelectedUserPostPopup from './SelectedUserPostPopup';
 
 const BASE_URL = import.meta.env.VITE_STRANGERS_THINGS_BASE_API
 
@@ -89,7 +90,7 @@ function UserPosts({query}) {
                     }
                 </tbody>
             </table>
-            <PostDetails 
+            <SelectedUserPostPopup 
                 trigger={trigger} 
                 setTrigger={setTrigger} 
                 id={selId} 
@@ -106,26 +107,3 @@ function UserPosts({query}) {
 }
 
 export default UserPosts
-
-function PostDetails({trigger, setTrigger, id, title, description, price, deliver, created, updated, location}){
-    return (
-        trigger?
-        <div className="popup" onClick={()=>setTrigger(false)}>
-            {/* <Link to="">  WE CAN TURN THIS INTO A LINK TO THE FULL POST PAGE THAT HAS MESSAGING OPTIONS ETC.*/}
-            <div className="popup-inner popup-neutral">
-                
-                Popup with all the post details! 
-                {id}   .
-                {title}  . 
-                {description}   .
-                {price}   .
-                {deliver}   .
-                {created}   .
-                {updated}   .
-                {location}
-            </div>
-            {/* </Link> */}
-
-        </div>:null
-    )
-}
